add video to the home add sass file for some files
