+++ src/App.js
@@ -6,9 +6,10 @@ import Cookies from "js-cookie";
 import { useDispatch } from "react-redux";
 import { getUser } from "./actions/user.actions";
 import "../node_modules/bootstrap/dist/css/bootstrap.min.css";
-import Footer from "./components/Footer";
+import Footer from "./components/Footer/Footer";
 import { useSetAtom } from "jotai";
 import { isLoggedAtom } from "./stores/user";
+import Navbar from "./components/Navbar/Navbar";
 
 const App = () => {
   const setIsLogged = useSetAtom(isLoggedAtom);
@@ -38,7 +39,8 @@ const App = () => {
 
   return (
     <UidContext.Provider value={uid}>
-      <Routes />
+      <Navbar/>     
+        <Routes />
       <Footer />
     </UidContext.Provider>
   );
