+++ src/components/BackgroundVideo/BackgroundVideo.scss
@@ -1,0 +1,65 @@
+#video-fond {
+  overflow: hidden;
+  position: fixed;
+  top: 128px;
+  right: 0;
+  bottom: 0;
+  left: 0;
+}
+#video-fond > video {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+}
+/* 1. Pas de support d'object-fit */
+@media (min-aspect-ratio: 16/9) {
+  #video-fond > video {
+    height: 300%;
+    top: -100%;
+  }
+}
+@media (max-aspect-ratio: 16/9) {
+  #video-fond > video {
+    width: 300%;
+    left: -100%;
+  }
+}
+/* 2. En cas de support d'object-fit, écrase les règles en (1) */
+@supports (object-fit: cover) {
+  #video-fond > video {
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    object-fit: cover;
+  }
+}
+
+// .video {
+//   background-color: black;
+//   background-image: "";
+//   background-position: center center;
+//   background-size: contain;
+// }
+
+#mavideo {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  object-fit: cover;
+}
+
+// .video {
+//   position: "fixed";
+//   width: "100%";
+//   left: "50%";
+//   top: "50%";
+//   height: "100%";
+//   object-fit: "cover";
+//   transform: "translate(-50%, -50%)";
+//   z-index: "-100";
+// }
