+++ src/components/Footer/Footer.scss
@@ -1,0 +1,3 @@
+#container-footer{
+    background-color: "";
+}
