+++ src/components/Navbar/Navbar.scss
@@ -1,0 +1,3 @@
+#container-header{
+    background-color: "";
+}
