+++ src/components/Routes/index.js
@@ -1,8 +1,7 @@
 import React from "react";
 import { Routes, Route } from "react-router-dom";
-import Home from "../../pages/Home";
+import Home from "../../pages/Home/Home";
 import Profil from "../../pages/Profil";
-import Navbar from "../Navbar";
 import Team from "../../pages/Team"
 import CreateProject from "../../pages/CreateProject"
 import Project from "../../pages/Project"
@@ -10,7 +9,6 @@ import Project from "../../pages/Project"
 const index = () => {
   return (
     <div>
-      <Navbar/>
         <Routes>
           <Route path="/" element={<Home />} />
           <Route path="/team" element={<Team />} />
