+++ src/pages/CreateProject.js
@@ -1,4 +1,7 @@
 import React from 'react';
+import AddPost from "../components/AddPost/AddPost";
+import NewPostForm from "../components/Post/NewPostForm";
+import Log from "../components/Log";
 
 const CreateProject = () => {
     return (
