+++ src/pages/Home/Home.scss
@@ -1,0 +1,3 @@
+.ThreadHome{
+    padding-top: 800px;
+}
