+++ src/styles/_base/_main.scss
@@ -18,30 +18,6 @@ body {
   color: white;  /* ############ 12 ############ */
 }
 
-// li {
-//   list-style-type: none;
-// }
-
-// a {
-//   text-decoration: none;
-//   color: $color-3;
-//   cursor: pointer;
-// }
-
-a:visited {
-  color: $color-3;
-}
-
-.btn-danger {
-  background: #dc3545;
-}
-.btn-success {
-  background: #28a745;
-}
-.btn-primary {
-  background: #007BFF;
-}
-
 svg {
   width: 70px;
   height: 70px;
